+++ Assets/Editor Toolbox/Editor/Utilities/ToolboxStyleUtility.cs
@@ -1,0 +1,7 @@
+﻿namespace Toolbox.Editor
+{
+    internal static class ToolboxStyleUtility
+    {
+        //this class will be used to handle internal style properties (in future realeses)
+    }
+}