+++ Assets/Editor Toolbox/CHANGELOG.md
@@ -1,5 +1,15 @@
-## 0.5.5.1 [09.01.2020]
+## 0.5.6 [16.01.2020]
+
+### Added:
+- Possibility to customize folder icons positions (check ToolboxEditorProject)
+
+### Changed:
+- Fix footer position of ReorderableList in Unity 2019.3+
+- Fix handle position of ReorderableList in Unity 2019.3+
+- Fix custom folders scalling 
+
+## 0.5.5 [09.01.2020]
 
 ### Added:
 - Tooltip for custom folders
-- Possibility to add drawers for hierarchy overlay from external classes (check ToolboxEditorHierarchy)
+- Possibility to add drawers to hierarchy overlay from external classes (check ToolboxEditorHierarchy)